requirement.txt
torch>=1.0.0
torchvision
torchnet
Pillow
numpy
scikit-image
thop